Created participant related classes. Modified entity relation types. Added some controller methods.

+++ src/main/java/com/sda/model/quizzes/Answer.java
@@ -9,11 +9,12 @@ public class Answer {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int answerId;
+    private Long answerId;
 
 //    @NotBlank(message = "Please write an answer")
     private String answerStatement;
 
+//    @Column(columnDefinition="tinyint(1) default 1")
     private boolean isAvailable;
 
 }
